Moved run model code related to Racing to a Model Concern
